pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "ninepatch"
version = "0.2.0"
authors = [
  { name="Thomas Schüßler", email="vindolin@gmail.com" },
]
dependencies = [
    "pillow",
]
description = "Slice and scale 9-patch images"
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

[project.urls]
"Homepage" = "https://github.com/vindolin/ninepatch"

[project.scripts]
ninepatch = 'ninepatch.commands.ninepatch:main'
ninepatch-viewer = 'ninepatch.commands.ninepatch_viewer:main'
